Refactoring: Core Data part 10 (Int32, isEqual)
- Using Int32 only in Core Data wrappers
- Entity wrapper check isEqual via managedObject

+++ Amperfy/Api/Ampache/ArtistParserDelegate.swift
@@ -23,7 +23,7 @@ class ArtistParserDelegate: GenericXmlLibParser {
                 artistBuffer = libraryStorage.createArtist()
                 artistBuffer?.syncInfo = syncWave
                 artistBuffer?.id = artistId
-                artistBuffer?.artwork?.url = ampacheUrlCreator.getArtUrlString(forArtistId: Int32(artistId))
+                artistBuffer?.artwork?.url = ampacheUrlCreator.getArtUrlString(forArtistId: artistId)
             }
 		}    
     }
